Add flock implementation
This is an implementation of flock(), pulled from the OpenSSH sources.

While AIX ships an flock implementation, it's in libbsd, which is
troublesome for many portable applications to use since it has
conflicting definitions of timezone. This can cause libc functions
gmtime, localtime, etc to fail in interesting ways.

Instead, we provide a definition in libutil, which does not conflict
with the version in libbsd so applications can use flock() without these
problems.