Fix backtrace with signal handlers interrupting (#4)
When an application (in my case, Mono), tries to get a backtrace
from a signal handler, it'll either stop because the backchain is
NULL (PASE) or fault again in the handler because it's some other
nonsensical value (AIX). This applies a heuristic to determine if
the frame is a signal handler, and if so, resets the stack pointer
to a location where the true backchain in the handler frame is
stored. The displacements used and idea of looking below the .text
base are inspired by how GDB attempts to handle this situation.

This commit fixes the issue and allows application signal handlers
to see the full backtrace, from before the signal was caught.