Don't compile using header wrappers
We have header wrappers for stdio.h and file.h which exist so they can
be installed and applications can get them seamlessly.

This causes problems, however, when the header wrapper in the build root
include_next's the same header wrapper that's been installed. Instead,
we should not use these header files to build, so they've been moved to
a separate "wrapper" directory.

Fixes #23